+++ include/plane_parallel.h
@@ -1,13 +1,13 @@
 #include <vector>
 
 
-std::vector<double> run_plane_parallel(const std::vector<double>& arr_z,
-                                      const std::vector<double>& arr_zh,
-                                      const std::vector<double>& arr_dz,
-                                      const std::vector<double>& arr_kext,
-                                      const std::vector<double>& arr_Batm,
-                                      const std::vector<double>& arr_Batmh,
-                                      double Bsfc,
+std::vector<float> run_plane_parallel(const std::vector<float>& arr_z,
+                                      const std::vector<float>& arr_zh,
+                                      const std::vector<float>& arr_dz,
+                                      const std::vector<float>& arr_kext,
+                                      const std::vector<float>& arr_Batm,
+                                      const std::vector<float>& arr_Batmh,
+                                      float Bsfc,
                                       int N_mu,
                                       bool print_EB,
                                       bool verbose);
